Add specific module to perform ASREP-roasting
