+++ lib/msf/core/exploit/remote/kerberos/client.rb
@@ -134,7 +134,7 @@ module Msf
             res
           end
 
-          # Sends a kerberos AS request and reads the response
+          # Sends a kerberos TGS request and reads the response
           #
           # @param opts [Hash]
           # @return [Rex::Proto::Kerberos::Model::KdcResponse]
@@ -285,7 +285,9 @@ module Msf
               ),
             )
 
-            initial_as_res = send_request_as(req: initial_as_req)
+            req_opts = {req: initial_as_req}
+            req_opts.update(options)
+            initial_as_res = send_request_as(req_opts)
 
             # If we receive an AS_REP response immediately, no-preauthentication was required and we can return immediately
             if initial_as_res.msg_type == Rex::Proto::Kerberos::Model::AS_REP
@@ -365,7 +367,9 @@ module Msf
                 )
               )
 
-              preauth_as_res = send_request_as(req: preauth_as_req)
+              req_opts = {req: preauth_as_req}
+              req_opts.update(options)
+              preauth_as_res = send_request_as(req_opts)
 
               # If we've succeeded - break out of trying ciphers
               break if preauth_as_res.msg_type == Rex::Proto::Kerberos::Model::AS_REP
